+++ sdk/resourcemanager/network/armnetwork/models.go
@@ -832,6 +832,9 @@ type ApplicationGatewayFrontendIPConfigurationPropertiesFormat struct {
 	// Reference to the subnet resource.
 	Subnet *SubResource
 
+	// READ-ONLY; Reference to the application gateway http listeners.
+	HTTPListeners *SubResource
+
 	// READ-ONLY; The provisioning state of the frontend IP configuration resource.
 	ProvisioningState *ProvisioningState
 }
@@ -1338,8 +1341,8 @@ type ApplicationGatewayProbePropertiesFormat struct {
 	PickHostNameFromBackendSettings *bool
 
 	// Custom port which will be used for probing the backend servers. The valid value ranges from 1 to 65535. In case not set,
-	// port from http settings will be used. This property is valid for Standardv2 and
-	// WAFv2 only.
+	// port from http settings will be used. This property is valid for Basic,
+	// Standardv2 and WAFv2 only.
 	Port *int32
 
 	// The protocol used for the probe.
@@ -2819,6 +2822,9 @@ type BackendAddressPoolPropertiesFormat struct {
 	// The location of the backend address pool.
 	Location *string
 
+	// Backend address synchronous mode for the backend pool
+	SyncMode *SyncMode
+
 	// An array of gateway load balancer tunnel interfaces.
 	TunnelInterfaces []*GatewayLoadBalancerTunnelInterface
 
@@ -3309,6 +3315,12 @@ type ConnectionMonitorDestination struct {
 	ResourceID *string
 }
 
+// ConnectionMonitorEndPointLocationDetails - Connection monitor endpoint location details.
+type ConnectionMonitorEndPointLocationDetails struct {
+	// Region for connection monitor endpoint.
+	Region *string
+}
+
 // ConnectionMonitorEndpoint - Describes the connection monitor endpoint.
 type ConnectionMonitorEndpoint struct {
 	// REQUIRED; The name of the connection monitor endpoint.
@@ -3323,11 +3335,17 @@ type ConnectionMonitorEndpoint struct {
 	// Filter for sub-items within the endpoint.
 	Filter *ConnectionMonitorEndpointFilter
 
+	// Location details for connection monitor endpoint.
+	LocationDetails *ConnectionMonitorEndPointLocationDetails
+
 	// Resource ID of the connection monitor endpoint.
 	ResourceID *string
 
 	// Endpoint scope.
 	Scope *ConnectionMonitorEndpointScope
+
+	// Subscription ID for connection monitor endpoint.
+	SubscriptionID *string
 
 	// The endpoint type.
 	Type *EndpointType
@@ -8322,6 +8340,18 @@ type MetricSpecification struct {
 
 	// Units the metric to be displayed in.
 	Unit *string
+}
+
+// MigrateLoadBalancerToIPBasedRequest - The request for a migrateToIpBased API.
+type MigrateLoadBalancerToIPBasedRequest struct {
+	// A list of pool names that should be migrated from Nic based to IP based pool
+	Pools []*string
+}
+
+// MigratedPools - The response for a migrateToIpBased API.
+type MigratedPools struct {
+	// A list of pools migrated from Nic based to IP based pool
+	MigratedPools []*string
 }
 
 // NatGateway - Nat Gateway resource.